imagediff/requirements.txt
Flask==3.0.3
Pillow==12.1.1
